# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BE-Smart API Documentation",
    "version": "1.0.0",
    "description": "API Documentation untuk E-Learning SMarT — Sistem Manajemen Sekolah. Dokumentasi ini mencakup semua endpoint yang tersedia beserta schema request/response.",
    "contact": {
      "name": "BE-Smart Team"
    }
  },
  "servers": [
    {
      "url": "https://localhost:3001",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Masukkan token JWT yang didapat dari proses login"
      }
    },
    "schemas": {
...
      "RegisterStudentRequest": {
        "type": "object",
        "required": [
          "username",
          "email",
          "phoneNumber",
          "password",
          "confirmPassword",
          "nik_ktp",
          "role",
          "nisn",
          "fatherName",
          "motherName",
          "parentPhone"
        ],
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "STUDENT"
            ],
            "example": "STUDENT"
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "example": "Ahmad Siswa"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "ahmad@student.example.com"
          },
          "phoneNumber": {
            "type": "string",
            "pattern": "^08[0-9]{8,11}$",
            "example": "081234567890"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "password123"
          },
          "confirmPassword": {
            "type": "string",
            "example": "password123"
          },
          "nik_ktp": {
            "type": "string",
            "pattern": "^\\d{16}$",
            "example": "3201234567890001"
          },
          "nisn": {
            "type": "string",
            "pattern": "^\\d{10}$",
            "example": "0012345678"
          },
          "fatherName": {
            "type": "string",
            "minLength": 3,
            "example": "Budi Santoso"
          },
          "motherName": {
            "type": "string",
            "minLength": 3,
            "example": "Siti Rahayu"
          },
          "parentPhone": {
            "type": "string",
            "pattern": "^08[0-9]{8,11}$",
            "example": "081298765432"
          }
        }
      },
      "RegisterTeacherRequest": {
        "type": "object",
        "required": [
          "username",
          "email",
          "phoneNumber",
          "password",
          "confirmPassword",
          "nik_ktp",
          "role",
          "specialization"
        ],
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "TEACHER"
            ],
            "example": "TEACHER"
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "example": "Pak Guru"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "guru@teacher.example.com"
          },
          "phoneNumber": {
            "type": "string",
            "pattern": "^08[0-9]{8,11}$",
            "example": "081234567890"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "password123"
          },
          "confirmPassword": {
            "type": "string",
            "example": "password123"
          },
          "nik_ktp": {
            "type": "string",
            "pattern": "^\\d{16}$",
            "example": "3201234567890002"
          },
          "nuptk": {
            "type": "string",
            "pattern": "^\\d{16}$",
            "example": "1234567890123456",
            "description": "Opsional — NUPTK 16 digit"
          },
          "nip": {
            "type": "string",
            "pattern": "^\\d{18}$",
            "example": "198501012010011001",
            "description": "Opsional — NIP 18 digit"
          },
          "specialization": {
            "type": "string",
            "minLength": 3,
            "example": "Matematika"
          },
          "educationLevel": {
            "type": "string",
            "enum": [
              "D3",
              "D4",
              "S1",
              "S2",
              "S3"
            ],
            "example": "S1",
            "description": "Opsional — Jenjang pendidikan terakhir"
          },
          "documents": {
            "type": "object",
            "properties": {
              "cv": {
                "type": "string",
                "example": "https://example.com/cv.pdf"
              },
              "certificates": {
                "type": "array",
                "items": {
                  "type": "string",
                  "example": "https://example.com/certificate1.pdf"
                }
              }
            },
            "required": [
              "cv",
              "certificates"
            ]
          }
        }
      },
      "RegisterStaffRequest": {
        "type": "object",
        "required": [
          "username",
          "email",
          "phoneNumber",
          "password",
          "confirmPassword",
          "nik_ktp",
          "role",
          "department"
        ],
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "STAFF"
            ],
            "example": "STAFF"
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "example": "Admin Staff"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "staff@school.example.com"
          },
          "phoneNumber": {
            "type": "string",
            "pattern": "^08[0-9]{8,11}$",
            "example": "081234567890"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "password123"
          },
          "confirmPassword": {
            "type": "string",
            "example": "password123"
          },
          "nik_ktp": {
            "type": "string",
            "pattern": "^\\d{16}$",
            "example": "3201234567890003"
          },
          "employeeId": {
            "type": "string",
            "example": "EMP-001",
            "description": "Opsional — ID karyawan"
          },
          "department": {
            "type": "string",
            "enum": [
              "ADMINISTRATION",
              "CLEANING_SERVICE",
              "COOPERATIVE",
              "FINANCE",
              "IT",
              "LABORATORY",
              "LIBRARY",
              "SECURITY"
            ],
            "example": "IT"
          },
          "employmentStatus": {
            "type": "string",
            "enum": [
              "CONTRACT",
              "HONORARY",
              "INTERN",
              "PERMANENT"
            ],
            "example": "PERMANENT",
            "description": "Opsional — Status kepegawaian"
          }
        }
      }
    }
  },
...
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register akun baru",
        "description": "Mendaftarkan user baru berdasarkan role (STUDENT, TEACHER, atau STAFF).\nSetiap role memiliki field tambahan yang berbeda.\nSetelah registrasi, user harus menunggu approval dari admin.\n",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/RegisterStudentRequest"
                  },
                  {
                    "$ref": "#/components/schemas/RegisterTeacherRequest"
                  },
                  {
                    "$ref": "#/components/schemas/RegisterStaffRequest"
                  }
                ]
              },
              "examples": {
                "student": {
                  "summary": "Register sebagai Siswa",
                  "value": {
                    "role": "STUDENT",
                    "username": "Ahmad Siswa",
                    "email": "ahmad@student.example.com",
                    "phoneNumber": "081234567890",
                    "password": "password123",
                    "confirmPassword": "password123",
                    "nik_ktp": "3201234567890001",
                    "nisn": "0012345678",
                    "fatherName": "Budi Santoso",
                    "motherName": "Siti Rahayu",
                    "parentPhone": "081298765432"
                  }
                },
                "teacher": {
                  "summary": "Register sebagai Guru",
                  "value": {
                    "role": "TEACHER",
                    "username": "Pak Guru",
                    "email": "guru@teacher.example.com",
                    "phoneNumber": "081234567890",
                    "password": "password123",
                    "confirmPassword": "password123",
                    "nik_ktp": "3201234567890002",
                    "nuptk": "1234567890123456",
                    "nip": "198501012010011001",
                    "specialization": "Matematika",
                    "educationLevel": "S1",
                    "documents": {
                      "cv": "https://example.com/cv.pdf",
                      "certificates": [
                        "https://example.com/certificate1.pdf",
                        "https://example.com/certificate2.pdf"
                      ]
                    }
                  }
                },
                "staff": {
                  "summary": "Register sebagai Staff",
                  "value": {
                    "role": "STAFF",
                    "username": "Admin Staff",
                    "email": "staff@school.example.com",
                    "phoneNumber": "081234567890",
                    "password": "password123",
                    "confirmPassword": "password123",
                    "nik_ktp": "3201234567890003",
                    "department": "IT"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registrasi berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "meta": {
                      "$ref": "#/components/schemas/MetaResponse"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        },
                        "profile": {
                          "oneOf": [
                            {
                              "$ref": "#/components/schemas/StudentProfile"
                            },
                            {
                              "$ref": "#/components/schemas/TeacherProfile"
                            }
                          ]
                        }
                      }
                    }
                  }
                },
                "example": {
                  "meta": {
                    "status": 200,
                    "message": "Registration successful. Please wait for admin approval."
                  },
                  "data": {
                    "user": {
                      "_id": "64f1a2b3c4d5e6f7a8b9c0d1",
                      "username": "Ahmad Siswa",
                      "email": "ahmad@student.example.com",
                      "roles": [
                        "STUDENT"
                      ],
                      "status": "PENDING",
                      "isApprove": "NOT_APPROVE"
                    },
                    "profile": {
                      "_id": "64f1a2b3c4d5e6f7a8b9c0d2",
                      "userId": "64f1a2b3c4d5e6f7a8b9c0d1",
                      "fatherName": "Budi Santoso",
                      "motherName": "Siti Rahayu"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                },
                "example": {
                  "meta": {
                    "status": 400,
                    "message": "Email tidak valid"
                  },
                  "data": {
                    "email": "Email tidak valid"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error / Email sudah terdaftar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "meta": {
                    "status": 500,
                    "message": "Registration failed, problem: Email Sudah terdaftar"
                  },
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login akun",
        "description": "Login menggunakan email, username, atau NIK dan password.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "identifier": "ahmad@student.example.com",
                "password": "password123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "meta": {
                    "status": 200,
                    "message": "Login success!"
                  },
                  "data": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Login gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Ambil informasi pengguna saat ini",
        "description": "Mengambil data user dan profil berdasarkan token JWT yang valid.",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Data user berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "meta": {
                    "status": 200,
                    "message": "User data retrieved successfully"
                  },
                  "data": [
                    {
                      "_id": "64f1a2b3c4d5e6f7a8b9c0d1",
                      "username": "Ahmad Siswa",
                      "email": "ahmad@student.example.com",
                      "roles": [
                        "STUDENT"
                      ],
                      "status": "ACTIVE",
                      "isApprove": "APPROVED"
                    },
                    {
                      "userId": "64f1a2b3c4d5e6f7a8b9c0d1",
                      "studentId": "0012345678",
                      "fatherName": "Budi Santoso",
                      "motherName": "Siti Rahayu"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / token tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Gagal mengambil data user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/activation": {
      "post": {
        "summary": "Aktivasi akun pengguna",
        "description": "Mengaktifkan akun setelah admin menyetujui user dengan activation code.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivationRequest"
              },
              "example": {
                "code": "activation-code-123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Aktivasi berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "meta": {
                    "status": 200,
                    "message": "Activation successful"
                  },
                  "data": [
                    {
                      "_id": "64f1a2b3c4d5e6f7a8b9c0d1",
                      "username": "Ahmad Siswa",
                      "email": "ahmad@student.example.com",
                      "status": "ACTIVE"
                    },
                    {
                      "_id": "64f1a2b3c4d5e6f7a8b9c0d2",
                      "userId": "64f1a2b3c4d5e6f7a8b9c0d1",
                      "status": "ACTIVE"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "User tidak disetujui atau kode tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Aktivasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/approve/{id}": {
      "patch": {
        "summary": "Setujui atau ubah status approval akun pengguna",
        "description": "Endpoint ini hanya dapat diakses oleh guru yang memiliki posisi Kepala Sekolah atau Wakil Kepala Sekolah. Identitas admin diambil dari token JWT yang sedang login.",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID pengguna target yang akan diubah status approval-nya"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "isApprove"
                ],
                "properties": {
                  "isApprove": {
                    "type": "string",
                    "enum": [
                      "APPROVED",
                      "PENDING",
                      "PENDING_PAYMENT",
                      "DAPODIK_ISSUE",
                      "DATA_ISSUE"
                    ],
                    "example": "APPROVED"
                  },
                  "targetUserId": {
                    "type": "string",
                    "description": "ID pengguna target yang akan diubah status approval-nya. Jika tidak dikirim, parameter path id akan dipakai.",
                    "example": "64f1a2b3c4d5e6f7a8b9c0d1"
                  },
                  "approvedAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-08-07T10:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status approval berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "meta": {
                    "status": 200,
                    "message": "User approval status updated successfully"
                  },
                  "data": {
                    "_id": "64f1a2b3c4d5e6f7a8b9c0d1",
                    "username": "Ahmad Siswa",
                    "email": "ahmad@student.example.com",
                    "isApprove": "APPROVED",
                    "approvedByUser": "admin",
                    "approvedAt": "2026-08-07T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Tidak memiliki akses / token tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Gagal memperbarui approval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}